Moved wallets over to @sei-js/core, updated chain suggest config.
