+++ packages/react/src/lib/components/WalletSelectModal/types.ts
@@ -1,4 +1,4 @@
-import { SeiWallet } from '../../provider';
+import { SeiWallet } from '@sei-js/core';
 
 export type WalletSelectModalProps = {
 	wallets?: SeiWallet[];
