--- packages/react/src/lib/config/supportedWallets.ts
@@ -1,84 +1,0 @@
-import { SeiWallet } from '../provider';
-
-export const FIN_WALLET: SeiWallet = {
-	getAccounts: async (chainId) => {
-		const offlineSigner = await window?.['fin']?.getOfflineSignerAuto(chainId);
-		return offlineSigner?.getAccounts() || [];
-	},
-	connect: async (chainId) => await window?.['fin']?.enable(chainId),
-	disconnect: async (chainId) => await window?.['fin']?.disable(chainId),
-	getOfflineSigner: async (chainId) => window?.['fin']?.getOfflineSignerAuto(chainId),
-	signArbitrary: async (chainId, signer, message) => window?.['fin']?.signArbitrary(chainId, signer, message),
-	walletInfo: {
-		windowKey: 'fin',
-		name: 'Fin',
-		website: 'https://chrome.google.com/webstore/detail/fin-wallet-for-sei/dbgnhckhnppddckangcjbkjnlddbjkna',
-		icon: 'https://sei-js-assets.s3.us-west-2.amazonaws.com/fin.png'
-	}
-};
-
-export const COMPASS_WALLET: SeiWallet = {
-	getAccounts: async (chainId) => {
-		const offlineSigner = await window?.['compass']?.getOfflineSignerAuto(chainId);
-		return offlineSigner?.getAccounts() || [];
-	},
-	connect: async (chainId) => await window?.['compass']?.enable(chainId),
-	disconnect: async (chainId) => await window?.['compass']?.disable(chainId),
-	getOfflineSigner: async (chainId) => window?.['compass']?.getOfflineSignerAuto(chainId),
-	signArbitrary: async (chainId, signer, message) => window?.['compass']?.signArbitrary(chainId, signer, message),
-	walletInfo: {
-		windowKey: 'compass',
-		name: 'Compass',
-		website: 'https://chrome.google.com/webstore/detail/compass-wallet/anokgmphncpekkhclmingpimjmcooifb',
-		icon: 'https://sei-js-assets.s3.us-west-2.amazonaws.com/compass.png'
-	}
-};
-
-export const KEPLR_WALLET: SeiWallet = {
-	getAccounts: async (chainId) => {
-		const offlineSigner = await window?.['keplr']?.getOfflineSignerAuto(chainId);
-		return offlineSigner?.getAccounts() || [];
-	},
-	connect: async (chainId) => await window?.['keplr']?.enable(chainId),
-	disconnect: async (chainId) => await window?.['keplr']?.disable(chainId),
-	getOfflineSigner: async (chainId) => window?.['keplr']?.getOfflineSignerAuto(chainId),
-	signArbitrary: async (chainId, signer, message) => window?.['keplr']?.signArbitrary(chainId, signer, message),
-	walletInfo: {
-		windowKey: 'keplr',
-		name: 'Keplr',
-		website: 'https://www.keplr.app/download',
-		icon: 'https://sei-js-assets.s3.us-west-2.amazonaws.com/keplr.png'
-	}
-};
-
-export const LEAP_WALLET: SeiWallet = {
-	getAccounts: async (chainId) => {
-		const offlineSigner = await window?.['leap']?.getOfflineSignerAuto(chainId);
-		return offlineSigner?.getAccounts() || [];
-	},
-	connect: async (chainId) => await window?.['leap']?.enable(chainId),
-	disconnect: async (chainId) => await window?.['leap']?.disable(chainId),
-	getOfflineSigner: async (chainId) => window?.['leap']?.getOfflineSignerAuto(chainId),
-	signArbitrary: async (chainId, signer, message) => window?.['leap']?.signArbitrary(chainId, signer, message),
-	walletInfo: {
-		windowKey: 'leap',
-		name: 'Leap',
-		website: 'https://www.leapwallet.io/download',
-		icon: 'https://sei-js-assets.s3.us-west-2.amazonaws.com/leap.png'
-	}
-};
-
-export const findWalletByWindowKey = (windowKey: string): SeiWallet | undefined => {
-	switch (windowKey) {
-		case 'compass':
-			return COMPASS_WALLET;
-		case 'leap':
-			return LEAP_WALLET;
-		case 'keplr':
-			return KEPLR_WALLET;
-		case 'fin':
-			return FIN_WALLET;
-		default:
-			return undefined;
-	}
-};
