+++ packages/react/src/lib/provider/types.ts
@@ -1,7 +1,7 @@
 import { AccountData, OfflineSigner } from '@cosmjs/proto-signing';
 import { Dispatch, ReactNode, SetStateAction } from 'react';
 import { ChainConfiguration } from '../types';
-import { StdSignature } from '@cosmjs/amino';
+import { SeiWallet } from '@sei-js/core';
 
 export type WalletProvider = {
 	chainId: string;
@@ -22,21 +22,6 @@ export type WalletProvider = {
 };
 
 export type SupportedWalletInput = 'compass' | 'leap' | 'fin' | 'keplr' | SeiWallet;
-
-export interface SeiWallet {
-	walletInfo: {
-		windowKey: string;
-		name: string;
-		icon: string;
-		website: string;
-	};
-	getOfflineSigner: (chainId: string) => Promise<OfflineSigner | undefined>;
-	getAccounts: (chainId: string) => Promise<readonly AccountData[]>;
-	connect: (chainId: string) => Promise<void>;
-	disconnect: (chainId: string) => Promise<void>;
-	suggestChain?: (chainId: string) => void;
-	signArbitrary?: (chainId: string, signer: string, message: string) => Promise<StdSignature | undefined>;
-}
 
 export type SeiWalletProviderProps = {
 	children: ReactNode;
